Centralize build graph check sources
