+++ docs/COMPLETION_AUDIT.md
@@ -2728,6 +2728,9 @@ and do not assume Phase 4 is ready without evidence.
   keeping typed AST node definitions smaller.
 - Parser import declarations now live in `src/parser/import_declarations.rs`,
   preserving parser import coverage while keeping declaration dispatch smaller.
+- Build graph check-source typechecking now lives in
+  `src/cli/build_graph_execution.rs`, preserving `zen check build.zen` source
+  diagnostics while keeping build graph validation helpers together.
 - `Self` expression and statement validation traversal now lives in
   `src/typechecker/self_type_validation/expressions.rs`, preserving the same
   context diagnostics while keeping declaration task collection and type
