+++ src/cli/build_graph_execution.rs
@@ -189,6 +189,23 @@ pub(super) fn validate_build_graph_sources(base_dir: &Path, graph: &zen::build_g
     }
 }
 
+pub(super) fn check_build_graph_sources(base_dir: &Path, graph: &zen::build_graph::BuildGraph) {
+    let mut source_paths = BTreeSet::new();
+    for target in graph.targets() {
+        for source in target.sources() {
+            source_paths.insert(base_dir.join(source));
+        }
+    }
+
+    for source_path in source_paths {
+        let source_path = source_path.to_str().unwrap_or_else(|| {
+            eprintln!("error: non-utf8 source path: {}", source_path.display());
+            process::exit(1);
+        });
+        super::graph_frontend(source_path);
+    }
+}
+
 fn validate_non_executed_target_sources(
     base_dir: &Path,
     graph: &zen::build_graph::BuildGraph,
